Retrieve the number of database links from Europe PMC publication database

This function returns the number of EBI database links associated with a publication.

epmc_db_count(ext_id = NULL, data_src = "med")

Arguments

ext_id
character, publication identifier

data_src
character, data source, by default Pubmed/MedLine index will be searched.

Value

data.frame with counts for each database

Details

Europe PMC supports cross-references between literature and the following databases:

'CHEBI'

a database and ontology of chemical entities of biological interest http://www.ebi.ac.uk/chebi/

'CHEMBL'

a database of bioactive drug-like small molecules https://www.ebi.ac.uk/chembldb/

'EMBL'

now ENA, provides a comprehensive record of the world's nucleotide sequencing information http://www.ebi.ac.uk/ena/

'INTACT'

provides a freely available, open source database system and analysis tools for molecular interaction data http://www.ebi.ac.uk/intact/

'INTERPRO'

provides functional analysis of proteins by classifying them into families and predicting domains and important sites http://www.ebi.ac.uk/interpro/

'OMIM'

a comprehensive and authoritative compendium of human genes and genetic phenotypes http://www.ncbi.nlm.nih.gov/omim

'PDB'

European resource for the collection, organisation and dissemination of data on biological macromolecular structures http://www.ebi.ac.uk/pdbe/

'UNIPROT'

comprehensive and freely accessible resource of protein sequence and functional information http://www.uniprot.org/

Examples

  
# NOT RUN {
  epmc_db_count(ext_id = "10779411")
  epmc_db_count(ext_id = "PMC3245140", data_src = "PMC")

# }